Add controller method/ refactor callbaclk functions
